requirements.txt
Flask>=2.3.0
requests>=2.31.0
Pillow>=9.0.0
pandas>=2.0.0
openpyxl>=3.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
